requirements.txt
numpy==1.24.1
Pillow==9.3.0
Pillow==10.0.0
torch==2.0.1+cu118
torchvision==0.15.2+cu118
tqdm==4.65.0
